Refactored CafeMap and CafeListItem to separate files
Also moved all global constants inside classes as static constants.
